Add files via upload
Archived project with beginning basic functionality.  Can insert words into a list which is then mapped with original word as a key, and reversed word as a value.  A text block is inserted and every defined swear that is found in the text block is reversed.  Likewise, a text block can be inserted into a sister method that will normalize every reversed swear.